Start of destabilizing changes :-(
Removed my detour of using uuid's, implemented something using sessions,
like pj's original implementation :-)

Remved all caches bar player counter.

We now use player names to id users/players, this probably should
be a number but this leaves the model (and lookup) as-is. uuid is
no longer used. NOTE this does mean that "old" games in the database
cant really be used even though the model didn't change as the username
index for players is different.